refactor: split steps.ts, client.ts, and workflow.ts into focused modules
Phase 1 — @drej/core: steps.ts → steps/ directory with 8 files. buildStep() is now a thin router; each step type lives in its own module (sandbox, exec, file, snapshot, control-flow). Circular dependency on buildStep avoided by passing it as a BuildStepFn parameter to control-flow builders.

Phase 2 — drej SDK: extract DrejError, WorkflowRun, and option types into types.ts; extract makeStream() into stream.ts as a pure function with no class dependency.

Phase 3 — drej SDK: workflow.ts → builder/ directory (types, sandbox-step, workflow, index). Zero public API changes across all packages.
