+++ packages/sdks/typescript/src/index.ts
@@ -30,6 +30,6 @@ export type {
   WorkflowFailedHookInfo,
 } from "./client";
 
-export { workflow, WorkflowBuilder, SandboxStepBuilder, CodeLanguage } from "./workflow";
-export type { SandboxOpts, LoopItem } from "./workflow";
+export { workflow, WorkflowBuilder, SandboxStepBuilder, CodeLanguage } from "./builder/index";
+export type { SandboxOpts, LoopItem } from "./builder/index";
 export type { CodeContext } from "@drej/opensandbox";
